Some doc fixes
