+++ darts/explainability/__init__.py
@@ -2,4 +2,6 @@
 Explainability
 --------------
 """
+
+from darts.explainability.explainability_result import ExplainabilityResult
 from darts.explainability.shap_explainer import ShapExplainer
